fix project gallery filtering and art completion sync
